code structure adjusts
