+++ src/components/WelcomePage/StyledWelcomePage.js
